feat: homepage is back
