+++ src/components/OutfitMakerComponent.tsx
