+++ src/pages/Home.tsx
@@ -1,72 +1,28 @@
 "use client";
 
-import { Box, Button, VStack, HStack, Text } from "native-base";
+import { Box, Button, VStack, Text } from "native-base";
 import { useState } from "react";
-import ClosetComponent from "../components/ClosetComponent";
-import WardrobeComponent from "../components/OutfitMakerCompnent";
-import EcoAdviceComponent from "../components/EcoAdviceComponent";
+import MainContent from "../components/MainContent";
 
 export default function Home() {
-  // State to track the active component
+  // State to track if the user has entered
+  const [hasEntered, setHasEntered] = useState(false);
   const [activeComponent, setActiveComponent] = useState("Component1");
 
-  // Components to render conditionally
-  const renderComponent = () => 
-    activeComponent === "Component1" ? (
-      <ClosetComponent/>
-    ) : activeComponent === "Component2" ? (
-      <WardrobeComponent/>
-    ) : activeComponent === "Component3" ? (
-      <EcoAdviceComponent/>
-    ) : (
-      <Text>Component 1</Text>
-    );
-
   return (
-    <Box w="100%" h="100vh" bg="white" safeArea>
-      <VStack space={4} flex={1} justifyContent="center" alignItems="center">
-      {/* <Image
-          source={{
-            uri: '/pwa-512x512.png'
-          }}
-          alt="Inside My Closet"
-          size="2xl"
-          resizeMode="contain"
-        /> */}
-        {renderComponent()}
-      </VStack>
-
-      {/* Navbar at the bottom */}
-      <HStack
-        w="100%"
-        position="absolute"
-        bottom={0}
-        bg="blue.500"
-        justifyContent="space-around"
-        p={4}
-      >
-        <Button
-          onPress={() => setActiveComponent("Component1")}
-          colorScheme="light"
-          variant={activeComponent === "Component1" ? "solid" : "outline"}
-        >
-          Button 1
-        </Button>
-        <Button
-          onPress={() => setActiveComponent("Component2")}
-          colorScheme="light"
-          variant={activeComponent === "Component2" ? "solid" : "outline"}
-        >
-          Button 2
-        </Button>
-        <Button
-          onPress={() => setActiveComponent("Component3")}
-          colorScheme="light"
-          variant={activeComponent === "Component3" ? "solid" : "outline"}
-        >
-          Button 3
-        </Button>
-      </HStack>
+    <Box w="100%" h="100vh" bg="white" safeArea justifyContent="center" alignItems="center">
+      {hasEntered ? (
+        <MainContent activeComponent={activeComponent} setActiveComponent={setActiveComponent} />
+      ) : (
+        <VStack space={4} flex={1} justifyContent="center" alignItems="center">
+          <Text fontSize="2xl" mb={4}>
+            Welcome to Inside My Closet!
+          </Text>
+          <Button onPress={() => setHasEntered(true)} colorScheme="blue">
+            Enter
+          </Button>
+        </VStack>
+      )}
     </Box>
   );
 }
